format diffs and covered changes as table
