+++ tools/qontract_cli.py
@@ -5,7 +5,7 @@ import re
 import sys
 from contextlib import suppress
 from datetime import datetime
-from typing import Dict, Iterable, List, Mapping, Optional, Union
+from typing import Optional
 
 import click
 import reconcile.gitlab_housekeeping as glhk
@@ -37,12 +37,12 @@ from reconcile.utils.jjb_client import JJB
 from reconcile.utils.mr.labels import SAAS_FILE_UPDATE
 from reconcile.utils.oc import OC_Map
 from reconcile.utils.ocm import OCMMap
+from reconcile.utils.output import print_output
 from reconcile.utils.secret_reader import SecretReader
 from reconcile.utils.semver_helper import parse_semver
 from reconcile.utils.state import State
 from reconcile.utils.terraform_client import TerraformClient as Terraform
 from reconcile.prometheus_rules_tester import get_rule_files_from_jinja_test_template
-from tabulate import tabulate
 
 from tools.sre_checkpoints import full_name, get_latest_sre_checkpoints
 from tools.cli_commands.gpg_encrypt import GPGEncryptCommand, GPGEncryptCommandData
@@ -1204,56 +1204,6 @@ def app_interface_merge_history(ctx):
     merge_queue_data.sort(key=itemgetter("merged_at"), reverse=True)
     ctx.obj["options"]["sort"] = False  # do not sort
     print_output(ctx.obj["options"], merge_queue_data, columns)
-
-
-def print_output(
-    options: Mapping[str, Union[str, bool]],
-    content: List[Dict],
-    columns: Iterable[str] = (),
-):
-    if options["sort"]:
-        content.sort(key=lambda c: tuple(c.values()))
-    if options.get("to_string"):
-        for c in content:
-            for k, v in c.items():
-                c[k] = str(v)
-
-    output = options["output"]
-
-    if output == "table":
-        print_table(content, columns)
-    elif output == "md":
-        print_table(content, columns, table_format="github")
-    elif output == "json":
-        print(json.dumps(content))
-    elif output == "yaml":
-        print(yaml.dump(content))
-    else:
-        pass  # error
-
-
-def print_table(content, columns, table_format="simple"):
-    headers = [column.upper() for column in columns]
-    table_data = []
-    for item in content:
-        row_data = []
-        for column in columns:
-            # example: for column 'cluster.name'
-            # cell = item['cluster']['name']
-            cell = item
-            for token in column.split("."):
-                cell = cell.get(token) or {}
-            if cell == {}:
-                cell = ""
-            if isinstance(cell, list):
-                if table_format == "github":
-                    cell = "<br />".join(cell)
-                else:
-                    cell = "\n".join(cell)
-            row_data.append(cell)
-        table_data.append(row_data)
-
-    print(tabulate(table_data, headers=headers, tablefmt=table_format))
 
 
 @root.group()
